Fix general Identifiable IDs not being of long type

+++ Client/Trivia/Source/Codec/C2S/Request/Packets/JoinRoomRequest.cs
@@ -1,7 +1,7 @@
 namespace Trivia.Codec.C2S.Request.Packets;
 
 public readonly record struct JoinRoomRequest(
-    int RoomId
+    long RoomId
 ) : IProtocolRequest
 {
     public RequestCode Code { get; init; } = RequestCode.JoinRoom;
